Complete the auth flow

+++ app/controllers/api/v1/users_controller.rb
@@ -1,22 +1,12 @@
 class Api::V1::UsersController < Api::V1::ApplicationController
     def index
         @user = User.all
-        render json: @user, status: :ok
+        render json: @user.as_json, status: :ok
     end
-
-
-    # def create
-    #     @user = User.new(user_params)
-    #     if @user.save
-    #         render json: @user, status: :created
-    #     else
-    #         render json: { error: @user.errors.full_message }, status: :unprocessable_entity
-    #     end
-    # end
 
     def show
         @user = User.find(params[:id])
-        render json: @user, status: :ok
+        render json: @user.as_json, status: :ok
     end
 
     def update
@@ -29,19 +19,22 @@ class Api::V1::UsersController < Api::V1::ApplicationController
     end
 
     def delete
-        @user = User.find(params[:id])
+        @user = User.find_by(id: params[:id])
 
-        render json: { error: @user.errors.full_message }, status: :not_found unless @user
+        unless @user
+            render json: { error: "User not found" }, status: :not_found
+            return
+        end
 
         if @user.destroy
-            render json: @user, status: :ok
+            render json: { message: "User deleted successfully" }, status: :ok
         else
-            render json: { error: @user.errors.full_message }, status: :unprocessable_entity
+            render json: { error: @user.errors.full_messages }, status: :unprocessable_entity
         end
     end
 
     private
     def user_params
-        params.require(:user).permit(:email, :password)
+        params.require(:user).permit(:email, :first_name, :last_name, :country, :date_of_birth)
     end
 end
